requirements.txt
scipy==1.11.3
typer==0.9.0
sentence_transformers==2.2.2
accelerate==0.21.0
sentencepiece==0.1.99
pydantic==1.10.13
transformers==4.36.2
chromadb==0.3.23
colorama==0.4.6
requests==2.31.0
peft==0.7.1
bitsandbytes==0.41.3.post2
tiktoken==0.5.2
openai==1.6.1
scikit-learn==1.3.2
wandb==0.16.1
numpy==1.26.2
datasets==2.15.0
python-dotenv==1.0.1
selenium==4.17.2
scrapy==2.11.1
html2text==2020.1.16
tqdm==4.66.1
